+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ffr",
+  "name": "example-structure",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
