+++ src/index.tsx
@@ -1,10 +1,16 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { App } from "./App";
+import AngularInjectorProvider from "./config/AngularInjector";
+import App from "./App";
+import { BrowserRouter } from "react-router-dom";
 
 const root = ReactDOM.createRoot(document.getElementById("root")!);
 root.render(
   <React.StrictMode>
-    <App />
+    <AngularInjectorProvider>
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
+    </AngularInjectorProvider>
   </React.StrictMode>
 );
