+++ src/templates/guest/Guest.tsx
@@ -1,0 +1,9 @@
+function Guest() {
+  return (
+    <>
+      <div>Guest</div>
+    </>
+  );
+}
+
+export default Guest;
