+++ src/templates/web/Web.tsx
@@ -1,0 +1,9 @@
+function Web() {
+  return (
+    <>
+      <div>Web</div>
+    </>
+  );
+}
+
+export default Web;
